tasks/temp/aigi-detection/requirements.txt
albumentations
einops
efficientnet-pytorch
filterpy
fvcore
imageio
imgaug
imutils
kornia
numpy
opencv-python
pandas
Pillow
pyyaml
scikit-image
scikit-learn
scipy
seaborn
segmentation-models-pytorch
setuptools
simplejson
tensorboard
tensorboardX
torch
torchaudio
torchtoolbox
torchvision
timm
transformers
tqdm

